export daemon message type & constants
